Move sunrise sunset logic
Put in the GET request since I need the IP of the caller, and it works functionally there
