+++ .gitignore
@@ -48,3 +48,4 @@ override.tf.json
 # Ignore CLI configuration files
 .terraformrc
 terraform.rc
+.DS_Store
