+++ terraform/terraform.tf
@@ -1,10 +1,10 @@
 terraform {
-  required_version = "1.9.7"
+  required_version = "1.10.3"
 
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.80.0"
+      version = "5.82.2"
     }
   }
 }
